feat: add a few core aspects
the whole idea of the game, 'meowlike', is to rely on certain degree of emulation/simulation world.
the main /wist - rules/laws of the world are changeable over time.
that includes things from the very basics of physics, till the details of differ ~magical systems/rituals 'executable', 'objective', to the particular 'reality'.

such a system may allow to create real flexible, 'fun', bizarre game situations to survive and explore.
in theory at least